@@ -1,18 +1,18 @@
 name: git_hooks
 description: Using dart to prevents bad commit or push (git hooks, pre-commit/precommit, pre-push/prepush, post-merge/postmerge and all that stuff...).It is similar to husky.
-version: 0.1.6
+version: 0.1.3-nullsafety.0
 homepage: https://github.com/xuzhongpeng/git_hooks
 
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  path: ^1.7.0
-  yaml: ^2.2.1
-  meta: ^1.2.4
+  path: ^1.8.0
+  yaml: ^3.1.0
+
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.3.0
+  pedantic: ^1.11.0
+  test: ^1.17.12
 executables:
   git_hooks:
