@@ -3,6 +3,7 @@
 import 'dart:async';
 import 'dart:io' as io;
 
+// ignore_for_file: public_member_api_docs
 class Ansi {
   static bool get terminalSupportsAnsi {
     return io.stdout.supportsAnsiEscapes &&
@@ -53,12 +54,12 @@ class Ansi {
 /// standard status messages, trace level output, and indeterminate progress.
 abstract class Logger {
   /// Create a normal [Logger]; this logger will not display trace level output.
-  factory Logger.standard({Ansi ansi}) => StandardLogger(ansi: ansi);
+  factory Logger.standard({Ansi? ansi}) => StandardLogger(ansi: ansi);
 
   /// Create a [Logger] that will display trace level output.
   ///
   /// If [logTime] is `true`, this logger will display the time of the message.
-  factory Logger.verbose({Ansi ansi, bool logTime = true}) {
+  factory Logger.verbose({Ansi? ansi, bool logTime = true}) {
     return VerboseLogger(ansi: ansi, logTime: logTime);
   }
 
@@ -93,7 +94,7 @@ abstract class Progress {
   Duration get elapsed => _stopwatch.elapsed;
 
   /// Finish the indeterminate progress display.
-  void finish({String message, bool showTiming});
+  void finish({String? message, bool showTiming = false});
 
   /// Cancel the indeterminate progress display.
   void cancel();
@@ -102,21 +103,19 @@ abstract class Progress {
 ///
 class StandardLogger implements Logger {
   @override
-  Ansi ansi;
+  final Ansi ansi;
 
-  StandardLogger({this.ansi}) {
-    ansi ??= Ansi(Ansi.terminalSupportsAnsi);
-  }
+  StandardLogger({Ansi? ansi}) : ansi = ansi ?? Ansi(Ansi.terminalSupportsAnsi);
 
   @override
   bool get isVerbose => false;
 
-  Progress _currentProgress;
+  Progress? _currentProgress;
 
   @override
   void stderr(String message) {
     if (_currentProgress != null) {
-      var progress = _currentProgress;
+      var progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -127,7 +126,7 @@ class StandardLogger implements Logger {
   @override
   void stdout(String message) {
     if (_currentProgress != null) {
-      var progress = _currentProgress;
+      var progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -141,7 +140,7 @@ class StandardLogger implements Logger {
   @override
   Progress progress(String message) {
     if (_currentProgress != null) {
-      var progress = _currentProgress;
+      var progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -169,7 +168,7 @@ class SimpleProgress extends Progress {
   void cancel() {}
 
   @override
-  void finish({String message, bool showTiming}) {}
+  void finish({String? message, bool? showTiming}) {}
 }
 
 class AnsiProgress extends Progress {
@@ -178,7 +177,7 @@ class AnsiProgress extends Progress {
   final Ansi ansi;
 
   int _index = 0;
-  Timer _timer;
+  late Timer _timer;
 
   AnsiProgress(this.ansi, String message) : super(message) {
     io.stdout.write('${message}...  '.padRight(40));
@@ -200,17 +199,21 @@ class AnsiProgress extends Progress {
   }
 
   @override
-  void finish({String message, bool showTiming = false}) {
+  void finish({String? message, bool showTiming = false}) {
     if (_timer.isActive) {
       _timer.cancel();
-      _updateDisplay(isFinal: true, message: message, showTiming: showTiming);
+      _updateDisplay(
+        isFinal: true,
+        message: message,
+        showTiming: showTiming,
+      );
     }
   }
 
   void _updateDisplay(
       {bool isFinal = false,
       bool cancelled = false,
-      String message,
+      String? message,
       bool showTiming = false}) {
     var char = kAnimationItems[_index % kAnimationItems.length];
     if (isFinal || cancelled) {
@@ -233,14 +236,13 @@ class AnsiProgress extends Progress {
 
 class VerboseLogger implements Logger {
   @override
-  Ansi ansi;
-  bool logTime;
-  Stopwatch _timer;
+  final Ansi ansi;
+  final bool logTime;
+  late Stopwatch _timer;
 
-  VerboseLogger({this.ansi, this.logTime}) {
-    ansi ??= Ansi(Ansi.terminalSupportsAnsi);
-    logTime ??= false;
-
+  VerboseLogger({Ansi? ansi, bool? logTime})
+      : ansi = ansi ?? Ansi(Ansi.terminalSupportsAnsi),
+        logTime = logTime ?? false {
     _timer = Stopwatch()..start();
   }
 
